fix(infra): release the apply stack lock on every exit + rename secret-paths
Two aftercare items from the 2026-08-10 IAM v2 migration
(.todos/IAM_V2_MIGRATION_PLAN.md §6b):

- apply.ts wraps everything after lock acquisition in try/finally with a
  guarded release: a thrown error or aborted prompt (Ctrl-C raises
  ExitPromptError) previously stranded the control lock until expiry, which
  blocked a CI deploy mid-migration.
- vm-reader-secret.ts → secret-paths.ts: the v1 principal the filename
  referred to is gone since the tranche-1/2 strips; the module holds the
  path layout + v2 condition builders. Pure rename, 13 importers updated.
