+++ infra/lib/scaleway/secret-paths.ts
