+++ infra/resources/secrets.ts
@@ -1,7 +1,7 @@
 import type * as pulumi from '@pulumi/pulumi';
 import * as scaleway from '@pulumiverse/scaleway';
 import { type RuntimeSecretDefinition, type RuntimeSecretId, runtimeSecrets } from '../lib/runtime-secrets';
-import { secretPathFor } from '../lib/scaleway/vm-reader-secret';
+import { secretPathFor } from '../lib/scaleway/secret-paths';
 import { mode, naming, region, tags } from '../pulumi-context';
 import { configuredOrRandomSecret } from './configured-secret';
 import { derivedRuntimeSecretData } from './stores';
