+++ infra/resources/vm-iam.ts
@@ -10,7 +10,7 @@ import {
   SERVICE_SECRET_PERMISSION_SETS,
 } from '../lib/scaleway/permissions';
 import { principalNames } from '../lib/scaleway/principals';
-import { bootKeyCondition, serviceKeyCondition } from '../lib/scaleway/vm-reader-secret';
+import { bootKeyCondition, serviceKeyCondition } from '../lib/scaleway/secret-paths';
 import { deployedServices, secretScopeSlugs } from '../lib/services';
 import { mode, naming, organizationId, projectId, tags } from '../pulumi-context';
