+++ infra/tasks/provision-managed-key.ts
@@ -2,7 +2,7 @@ import type { ManagedKeyDefinition, MintedKeyField } from '../lib/managed-keys';
 import { type RuntimeSecretDefinition, runtimeSecrets } from '../lib/runtime-secrets';
 import { provisionScopedKey } from '../lib/scaleway/scaleway-iam';
 import { createSecretManagerClient } from '../lib/scaleway/scaleway-secret-manager';
-import { secretPathFor } from '../lib/scaleway/vm-reader-secret';
+import { secretPathFor } from '../lib/scaleway/secret-paths';
 import { checkMark } from '../lib/utils/cli-output';
 
 export interface ProvisionManagedKeyOptions {
