+++ infra/tasks/status.ts
@@ -9,7 +9,7 @@ import { deriveInfra } from '../lib/naming';
 import { operatorManagedRuntimeSecrets } from '../lib/runtime-secrets';
 import { resolveProjectId } from '../lib/scaleway/bootstrap-scw-env';
 import { createSecretManagerClient } from '../lib/scaleway/scaleway-secret-manager';
-import { secretManagerPath } from '../lib/scaleway/vm-reader-secret';
+import { secretManagerPath } from '../lib/scaleway/secret-paths';
 import { serviceEndpoints } from '../lib/services';
 import {
   detectComputeDeferred,
